@@ -16,25 +16,25 @@
 		"@inlang/paraglide-js": "1.11.2",
 		"@sveltejs/adapter-auto": "^3.2.4",
 		"@sveltejs/adapter-vercel": "^5.4.3",
-		"@sveltejs/kit": "^2.5.24",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
-		"@tailwindcss/typography": "^0.5.14",
-		"@types/eslint": "^8.56.11",
+		"@sveltejs/kit": "^2.5.26",
+		"@sveltejs/vite-plugin-svelte": "4.0.0-next.7",
+		"@tailwindcss/typography": "^0.5.15",
+		"@types/eslint": "^8.56.12",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.9.0",
+		"eslint": "^9.10.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.43.0",
 		"globals": "^15.9.0",
-		"postcss": "^8.4.41",
+		"postcss": "^8.4.45",
 		"prettier": "^3.3.3",
 		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.6",
 		"svelte": "5.0.0-next.172",
 		"svelte-check": "^3.8.6",
 		"tailwindcss": "^3.4.10",
-		"tslib": "^2.6.3",
+		"tslib": "^2.7.0",
 		"typescript-eslint": "8.0.0-alpha.39",
-		"vite": "^5.4.2",
+		"vite": "^5.4.3",
 		"zod": "^3.23.8"
 	},
 	"type": "module",
@@ -43,7 +43,7 @@
 		"@fontsource/jetbrains-mono": "^5.0.21",
 		"@inlang/paraglide-sveltekit": "0.11.0",
 		"@lucia-auth/adapter-prisma": "^4.0.1",
-		"@prisma/client": "^5.18.0",
+		"@prisma/client": "^5.19.1",
 		"@prisma/extension-accelerate": "^1.1.0",
 		"arctic": "^1.9.2",
 		"bits-ui": "^0.21.13",
