@@ -12,7 +12,7 @@
  - size: number = 24 - is the width and height of the svg
  - strokeWidth: number = 2 - is the thickness of the stroke
  - absoluteStrokeWidth: boolean = false - if true, the stroke width will be the same regardless of the size of the icon
- - classes: string (tailwindcss color format e.g. text-blue-500)
+ - CLASS: string (tailwindcss color format e.g. text-blue-500)
  - children: Snippet
 -->
 <Icon name="Sun" {...props}>
